Styling incident-lite and moving to incidents folder
