+++ app/js/components/incident/incidentIcon/_incident-icon.scss
