+++ app/js/components/incident/incidentSummary/incidentSummary.jade
@@ -1,33 +1,32 @@
-.row
-  .col-sm-3(ng-if='incidentCount > 0')
-    #countComparison.countComparison
-      .count
-        .num {{incidentCount}}
-        .desc(translate) Incidents
-      .count
-        .num {{incidentSystemCount}}
-        .desc(translate) Systems
-  .col-sm-9.border-left(ng-if='incidentCount > 0')
-    .icon-inline
-      i.fa.fa-bullseye
-      p An&nbsp;
-        strong(translate) incident&nbsp;
-        | is an issue that is currently or has already affected systems. {{incidentCount}} of 
-        span(translate, translate-n="ruleHits", translate-plural=" {{$count}} issues") {{ruleHits}} issue 
-        span(translate, translate-n="incidentCount", translate-plural="&nbsp;are INCIDENTS affecting") &nbsp;is an INCIDENT affecting 
-        span(translate, translate-n="incidentSystemCount", translate-plural="&nbsp;{{$count}} systems.") &nbsp;{{incidentSystemCount}} system. 
-        span(translate) &nbsp;It is imperative that these incidents be addressed to ensure optimal performance.
-      br
-      a.link.strong(translate, ui-sref="app.topic({id: 'incidents'})") View incidents
+.cell.counts(ng-if='incidentCount > 0')
+  .count
+    h4.num {{incidentCount}}
+    span(translate) Incidents
+  
+  .count
+    h4.num {{incidentSystemCount}}
+    span(translate) Systems
 
-  .col-md-12(ng-if='incidentCount === 0')
-    #countComparison.countComparison
-      .count
-        span.num.success 0
-        span.desc incidents
-  .col-md-12(ng-if='incidentCount === 0')
-    .icon-inline
-      i.fa.fa-bullseye
-      p(translate) An&nbsp;
-        strong incident&nbsp;
-        | is an issue that is currently or has already affected systems. You have 0 issues that are INCIDENTS.
+.cell.counts(ng-if='incidentCount === 0')
+  .count
+    h4.num.success 0
+    span(translate) incidents
+
+.cell.supporting-text
+  .icon-inline(ng-if='incidentCount > 0')
+    i.fa.fa-bullseye(aria-hidden="true")
+    p An&nbsp;
+      strong(translate) incident&nbsp;
+      | is an issue that is currently or has already affected systems. {{incidentCount}} of 
+      span(translate, translate-n="ruleHits", translate-plural=" {{$count}} issues") {{ruleHits}} issue 
+      span(translate, translate-n="incidentCount", translate-plural="&nbsp;are INCIDENTS affecting") &nbsp;is an INCIDENT affecting 
+      span(translate, translate-n="incidentSystemCount", translate-plural="&nbsp;{{$count}} systems.") &nbsp;{{incidentSystemCount}} system. 
+      span(translate) &nbsp;It is imperative that these incidents be addressed to ensure optimal performance.
+    br
+    a.link.strong(translate, ui-sref="app.topic({id: 'incidents'})") View incidents
+
+  .icon-inline(ng-if='incidentCount === 0')
+    i.fa.fa-bullseye(aria-hidden="true")
+    p(translate) An&nbsp;
+      strong incident&nbsp;
+      | is an issue that is currently or has already affected systems. You have 0 issues that are INCIDENTS.
