--- app/js/components/overview/widgets/incidentLite/incidentLite.jade
@@ -1,20 +1,0 @@
-.content
-  .row(ng-show='incidentCount !== undefined')
-    .col-sm-6(ng-if='incidentCount > 0')
-      #countComparison.countComparison
-        .count
-          .num {{incidentCount}}
-          .desc(translate) Incidents
-    .col-sm-6(ng-if='incidentCount > 0')
-      #countComparison.countComparison
-        .count
-          .num {{incidentSystemCount}}
-          .desc(translate) Systems
-
-    .col-sm-12(ng-if='incidentCount === 0')
-      #countComparison.countComparison
-        .count
-          span.num.success 0
-          span.desc incidents
-  .row(ng-show='incidentCount === undefined')
-    .spinner
