+++ app/styles/components/_grid-view.scss
@@ -7,7 +7,6 @@
 .grid-block {
 	margin-left: -15px;
 	margin-right: -15px;
-	@include rem('padding', 15px);
 }
 
 .grid-block + .grid-block {
