+++ app/styles/components/_icons.scss
@@ -15,7 +15,7 @@ i.fa.category-icon {
 }
 
 i.fa.fa-bullseye {
-    color: red;
+  color: $red;
 }
 
 img.icon {
