+++ app/styles/includes/_vars.scss
@@ -13,6 +13,8 @@ $lg: 1200px;
 $base-font-size: 16px;
 $base-color: #4c4c4c;
 $base-font-family: 'Overpass', Arial, Helvetica, sans-serif;
+$brand-gray: $base-color;
+
 
 // Global Settings
 //==================================================================*/
